+++ src/MultipleSelect.tsx
@@ -1,4 +1,4 @@
-import { Flex, Typography, useTheme } from "ingred-ui";
+import { Flex } from "ingred-ui";
 import React, {
   ChangeEvent,
   createRef,
@@ -8,9 +8,8 @@ import React, {
   useCallback,
   useState,
 } from "react";
-import styled from "styled-components";
 import { MultipleInput } from "./Input";
-import { Option } from "./Option";
+import { Menu } from "./utils/Menu/Menu";
 import { OptionType } from "./types";
 
 type Props = {
@@ -53,6 +52,7 @@ export const MultipleSelect = forwardRef<HTMLDivElement, Props>(
       onMouseDownOption,
       onClickSelectContainer,
       onChangeSelected,
+      onChangeValue,
       ...rest
     },
     ref
@@ -67,6 +67,7 @@ export const MultipleSelect = forwardRef<HTMLDivElement, Props>(
       const newValue = selected?.slice(0, -1) ?? null;
       setSelected(newValue);
       onChangeSelected && onChangeSelected(newValue);
+      onChangeValue && onChangeValue(newValue);
     };
 
     const handleRemoveValue = (
@@ -78,6 +79,7 @@ export const MultipleSelect = forwardRef<HTMLDivElement, Props>(
         selected?.filter((s: OptionType) => s.value !== value) ?? null;
       setSelected(newValue);
       onChangeSelected && onChangeSelected(newValue);
+      onChangeValue && onChangeValue(newValue);
     };
 
     const handleClickOption = useCallback(
@@ -87,6 +89,7 @@ export const MultipleSelect = forwardRef<HTMLDivElement, Props>(
         const newValue = [...(selected ?? []), data];
         setSelected(newValue);
         onChangeSelected && onChangeSelected(newValue);
+        onChangeValue && onChangeValue(newValue);
         onFocusInput && onFocusInput();
 
         onClearInput();
@@ -106,8 +109,9 @@ export const MultipleSelect = forwardRef<HTMLDivElement, Props>(
         return;
       }
 
-      setSelected(null);
-      onChangeSelected && onChangeSelected(null);
+      setSelected([]);
+      onChangeSelected && onChangeSelected([]);
+      onChangeValue && onChangeValue([]);
 
       event.preventDefault();
       if (event.type === "touchend" && onFocusInput) {
@@ -139,6 +143,7 @@ export const MultipleSelect = forwardRef<HTMLDivElement, Props>(
             handleRemoveValue={handleRemoveValue}
           />
           <Menu
+            isMulti
             ref={menuRef}
             selected={selected}
             options={options}
@@ -148,89 +153,6 @@ export const MultipleSelect = forwardRef<HTMLDivElement, Props>(
             onMouseDownOption={onMouseDownOption}
           />
         </Flex>
-      </Flex>
-    );
-  }
-);
-
-const MenuContainer = styled.div`
-  padding: ${({ theme }) => `${theme.spacing / 2}px 0`};
-  border-radius: ${({ theme }) => `0 0 ${theme.radius}px ${theme.radius}px`};
-  border: ${({ theme }) => `1px solid ${theme.palette.gray.main}`};
-  border-top: none;
-`;
-
-const EmptyOptions = () => {
-  const theme = useTheme();
-  return (
-    <Flex height="64px">
-      <Typography color={theme.palette.gray.main} align="center">
-        Not found
-      </Typography>
-    </Flex>
-  );
-};
-
-type MenuProps = {
-  selected: any | undefined;
-  menuIsOpen: boolean;
-  options: OptionType[];
-  optionsValue: OptionType[];
-  onClickOption: (data: OptionType) => void;
-  onMouseDownOption: (
-    event: MouseEvent<HTMLDivElement, globalThis.MouseEvent>
-  ) => void;
-};
-
-export const Menu = forwardRef<HTMLDivElement, MenuProps>(
-  (
-    {
-      options,
-      selected,
-      menuIsOpen,
-      optionsValue,
-      onClickOption,
-      onMouseDownOption,
-    },
-    ref
-  ) => {
-    const getMenuList = () => {
-      if (selected == null) {
-        return optionsValue.map(({ value, label }, index) => (
-          <Option
-            key={`${value}-${label}-${index}`}
-            value={value}
-            label={label}
-            onClickOption={onClickOption}
-            selected={selected?.value === value}
-          />
-        ));
-      }
-
-      if (selected.length === options.length) {
-        return <EmptyOptions />;
-      }
-
-      const values = selected.map(
-        ({ value }: { value: string | number }) => value
-      );
-
-      return optionsValue
-        .filter(({ value }) => values.indexOf(value) === -1)
-        .map(({ value, label }, index) => (
-          <Option
-            key={`${value}-${label}-${index}`}
-            value={value}
-            label={label}
-            onClickOption={onClickOption}
-            selected={selected?.value === value}
-          />
-        ));
-    };
-
-    return (
-      <Flex ref={ref} onMouseDown={onMouseDownOption}>
-        {menuIsOpen ? <MenuContainer>{getMenuList()}</MenuContainer> : null}
       </Flex>
     );
   }
